Put md5 function back into files
