+++ src/contacts.html
@@ -11,7 +11,7 @@
 		<link rel="stylesheet" href="./styles/reset.css" type="text/css">
 		<link rel="stylesheet" href="./styles/contacts.css" type="text/css">
 
-		<script src="./scripts/contacts.js" type="module"></script>
+		<script src="./scripts/contacts.js"></script>
 	</head>
 
 	<body>
