+++ src/index.html
@@ -11,7 +11,7 @@
 		<link rel="stylesheet" href="./styles/reset.css" type="text/css">
 		<link rel="stylesheet" href="./styles/index.css" type="text/css">
 
-		<script src="./scripts/index.js" type="module"></script>
+		<script src="./scripts/index.js"></script>
 	</head>
 
 	<body>
